+++ hutool-cron/src/test/java/cn/hutool/cron/pattern/CronPatternNextMatchTest.java
@@ -1,8 +1,10 @@
 package cn.hutool.cron.pattern;
 
 import cn.hutool.core.date.DateUtil;
+import cn.hutool.core.date.Week;
 import cn.hutool.core.lang.Console;
 import org.junit.Assert;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import java.util.Calendar;
@@ -18,5 +20,14 @@ public class CronPatternNextMatchTest {
 
 		Console.log(DateUtil.date(calendar));
 		Assert.assertTrue(pattern.match(calendar, true));
+	}
+
+	@Test
+	@Ignore
+	public void calendarTest(){
+		final Calendar ca = Calendar.getInstance();
+		ca.set(Calendar.DAY_OF_WEEK, Week.SATURDAY.getValue());
+
+		Console.log(DateUtil.date(ca));
 	}
 }