+++ hutool-json/src/test/java/cn/hutool/json/JSONObjectTest.java
@@ -289,7 +289,7 @@ public class JSONObjectTest {
 		userA.setDate(new Date());
 		userA.setSqs(CollectionUtil.newArrayList(new Seq(null), new Seq("seq2")));
 
-		JSONObject json = JSONUtil.parseObj(userA, false, true);
+		JSONObject json = JSONUtil.parseObj(userA, false);
 
 		Assert.assertTrue(json.containsKey("a"));
 		Assert.assertTrue(json.getJSONArray("sqs").getJSONObject(0).containsKey("seq"));
@@ -422,7 +422,6 @@ public class JSONObjectTest {
 	public void setDateFormatTest() {
 		JSONConfig jsonConfig = JSONConfig.create();
 		jsonConfig.setDateFormat("yyyy-MM-dd HH:mm:ss");
-		jsonConfig.setOrder(true);
 
 		JSONObject json = new JSONObject(jsonConfig);
 		json.append("date", DateUtil.parse("2020-06-05 11:16:11"));
@@ -435,7 +434,6 @@ public class JSONObjectTest {
 	public void setDateFormatTest2() {
 		JSONConfig jsonConfig = JSONConfig.create();
 		jsonConfig.setDateFormat("yyyy#MM#dd");
-		jsonConfig.setOrder(true);
 
 		Date date = DateUtil.parse("2020-06-05 11:16:11");
 		JSONObject json = new JSONObject(jsonConfig);
@@ -456,7 +454,6 @@ public class JSONObjectTest {
 	public void setCustomDateFormatTest() {
 		JSONConfig jsonConfig = JSONConfig.create();
 		jsonConfig.setDateFormat("#sss");
-		jsonConfig.setOrder(true);
 
 		Date date = DateUtil.parse("2020-06-05 11:16:11");
 		JSONObject json = new JSONObject(jsonConfig);
@@ -616,7 +613,7 @@ public class JSONObjectTest {
 
 	@Test
 	public void filterIncludeTest() {
-		JSONObject json1 = JSONUtil.createObj(JSONConfig.create().setOrder(true))
+		JSONObject json1 = JSONUtil.createObj(JSONConfig.create())
 				.set("a", "value1")
 				.set("b", "value2")
 				.set("c", "value3")
@@ -628,7 +625,7 @@ public class JSONObjectTest {
 
 	@Test
 	public void filterExcludeTest() {
-		JSONObject json1 = JSONUtil.createObj(JSONConfig.create().setOrder(true))
+		JSONObject json1 = JSONUtil.createObj(JSONConfig.create())
 				.set("a", "value1")
 				.set("b", "value2")
 				.set("c", "value3")
@@ -640,7 +637,7 @@ public class JSONObjectTest {
 
 	@Test
 	public void editTest() {
-		JSONObject json1 = JSONUtil.createObj(JSONConfig.create().setOrder(true))
+		JSONObject json1 = JSONUtil.createObj(JSONConfig.create())
 				.set("a", "value1")
 				.set("b", "value2")
 				.set("c", "value3")
@@ -660,7 +657,7 @@ public class JSONObjectTest {
 
 	@Test
 	public void toUnderLineCaseTest() {
-		JSONObject json1 = JSONUtil.createObj(JSONConfig.create().setOrder(true))
+		JSONObject json1 = JSONUtil.createObj(JSONConfig.create())
 				.set("aKey", "value1")
 				.set("bJob", "value2")
 				.set("cGood", "value3")
@@ -675,7 +672,7 @@ public class JSONObjectTest {
 
 	@Test
 	public void nullToEmptyTest() {
-		JSONObject json1 = JSONUtil.createObj(JSONConfig.create().setOrder(true).setIgnoreNullValue(false))
+		JSONObject json1 = JSONUtil.createObj(JSONConfig.create().setIgnoreNullValue(false))
 				.set("a", null)
 				.set("b", "value2");